Squashed delete method bug
Called deleteNode() on the in order successor without assigning the return value to anything (which is actually how the deletion is performed, by assigning the return value to the parent, effectively passing over the node being deleted). Instead, needed to reassign the right branch to the children of the node being deleted.